+++ EnhanceQoL/Settings/Loot.lua
@@ -38,6 +38,7 @@ end
 local cLoot = addon.SettingsLayout.rootGENERAL
 local expandable = addon.functions.SettingsCreateExpandableSection(cLoot, {
 	name = L["Loot"],
+	iconKey = "loot",
 	expanded = false,
 	colorizeTitle = false,
 	newTagID = "Loot",